refactor(server): share audit event recording helper
